pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.1",
    "inngest>=0.5.11",
    "llama-index-core>=0.14.8",
    "llama-index-llms-gemini>=0.6.1",
    "llama-index-readers-file>=0.5.4",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.15.1",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
]
